Improve exception handling in resetState
Ideally every separate thing being reset would be wrapped in its own try-catch
block, but that would make the code hard to maintain so I'm just wrapping the
entirety of resetState in a try-catch block.

The most important resets are placed at the beginning of the function to make
them the most likely to get executed before bailing out.