Add compatibility hack for FoamFix's jarDiscovererMemoryLeakFix
It turns out porting the patch to ASM didn't reliably fix the incompatibility.
It only worked when my transformer ran before FF's, but we have the same sort
order so it isn't guaranteed that this will be the case. For example, with
CoreTweaks+FoamFix+SpongeMixins an error still happened.

Now I hack FoamFix to not apply its patch, which should fully fix the issue
as long as FF's internals don't change.